feat: Complete Step1 DeFi System - Autonomous Trading with ASC
🚀 Major Features Implemented:
- Rule-based market detection (PracticalMarketDetector)
- Autonomous trading execution (TradingExecutor)
- ASC scheduling system (AutonomousScheduler)
- Real-time data aggregation (DataAggregator)
- Comprehensive risk management
- Production-ready error handling and logging

🔧 Technical Achievements:
- Full English codebase and documentation
- AssemblyScript compatibility optimizations
- Gas-efficient on-chain execution
- Multi-source data validation
- Thread-safe state management

🎯 AKINDO x Massa Buildathon Ready:
- Complete autonomous DeFi system
- Massa ASC integration
- Production-grade architecture
- Comprehensive README documentation

+++ src/deploy.ts
@@ -58,7 +58,7 @@ async function main() {
     for (let i = 0; i < 3; i++) {
       const threadArgs = new Args()
         .addU8(i)           // Thread ID
-        .addI64(1000 * (i + 1)); // Initial value
+        .addI64(BigInt(1000 * (i + 1))); // Initial value
       
       const thread = await deployContract('ObservationThread', threadArgs, i);
       observationThreads.push(thread);
